refactor: Restructure screenshot generation logic and update related imports for improved clarity and maintainability
