+++ apps/server/src/infra/db/index.ts
@@ -1,4 +1,5 @@
-import { PrismaClient } from "../../prisma/generated/client";
+import { PrismaClient } from "../../../prisma/generated/client";
+
 const prisma = new PrismaClient();
 
 export default prisma;
