--- apps/server/src/modules/generate/screenshot/route.ts
@@ -1,166 +1,0 @@
-import type { CustomerState } from "@polar-sh/sdk/models/components/customerstate.js";
-import Elysia, { t } from "elysia";
-import { Env } from "@/env/schema";
-import prisma from "@/infra/db";
-import { polarClient } from "@/lib/payments";
-import { putObject } from "@/lib/s3-client";
-import {
-  extractContentType,
-  sanitizeAndValidateUrl,
-} from "@/utils/url-helpers";
-
-function hasActiveSubscription(customerState: CustomerState) {
-  return customerState.activeSubscriptions?.some(
-    (sub) => sub.status === "active"
-  );
-}
-
-function hasMeterBalance(customerState: CustomerState) {
-  return customerState.activeMeters?.some((meter) => meter.balance > 0);
-}
-
-export const routeScreenshot = new Elysia().get(
-  "/generate-screenshot",
-  async (context) => {
-    // http://fyndx.io/generate/screenshot?url=https://social-image-maker.vercel.app/blog/article-1?utm_source=twitter
-    const { url: queryUrl } = context.query;
-    const sanitizedUrl = sanitizeAndValidateUrl(queryUrl);
-    // Check if the domain exists in the projects table
-    // If exists, get the user associated with the project
-    const project = await prisma.project.findUnique({
-      where: {
-        domain: sanitizedUrl.origin,
-      },
-      include: { user: true },
-    });
-
-    if (!project) {
-      return context.status(404, "Project not found for the given domain");
-    }
-    // 6. Check if the user is subscribed to a plan
-    // TODO: move external state to internal state on user table
-    const customerState = await polarClient.customers.getStateExternal({
-      externalId: project?.user.id,
-    });
-
-    const hasSubscription = hasActiveSubscription(customerState);
-
-    if (!hasSubscription) {
-      return context.status(403, "User does not have an active subscription");
-    }
-    // 6.1 check if the page url screenshot is in db
-    const screenshotUrl = `${sanitizedUrl.origin}${sanitizedUrl.pathname}`;
-    const existingImage = await prisma.generatedImage.findFirst({
-      where: {
-        projectId: project.id,
-        urlPath: screenshotUrl,
-      },
-    });
-
-    // if exists, return the image blob instead of url
-    if (existingImage) {
-      // TODO: Test this
-      return context.status(302, existingImage.imageUrl);
-    }
-
-    // if image does not exist, check if user has enough credits
-    const hasBalance = hasMeterBalance(customerState);
-
-    if (!hasBalance) {
-      return context.status(403, "User does not have enough credits");
-    }
-
-    // 6.1 Check if the page exists and gives 200
-    const checkPageExists = await fetch(screenshotUrl, {
-      method: "HEAD",
-    });
-
-    if (!checkPageExists.ok) {
-      return context.status(
-        checkPageExists.status,
-        `The requested page is not available (HTTP ${checkPageExists.status})`
-      );
-    }
-
-    const encodedUrl = encodeURIComponent(screenshotUrl);
-    const browserlessUrl = Env.BROWSERLESS_URL;
-    const browserlessToken = Env.BROWSERLESS_TOKEN;
-    if (!(browserlessUrl && browserlessToken)) {
-      return context.status(500, "Browserless configuration error");
-    }
-    const apiUrl = `${browserlessUrl}/screenshot?token=${browserlessToken}&stealth=true`;
-    const screenshotResponse = await fetch(apiUrl, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        url: screenshotUrl,
-        userAgent: {
-          userAgent:
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-        },
-        options: {
-          type: "jpeg",
-          quality: 100,
-        },
-        gotoOptions: {
-          waitUntil: "networkidle0",
-        },
-        viewport: {
-          width: 1200,
-          height: 630,
-          deviceScaleFactor: 2,
-        },
-      }),
-    });
-
-    if (screenshotResponse.ok) {
-      const contentType = extractContentType(screenshotResponse.headers);
-      const arrayBuffer = await screenshotResponse.arrayBuffer();
-      const buffer = Buffer.from(arrayBuffer);
-
-      const { success, error } = await putObject({
-        bucket: "og-images-staging",
-        key: `${encodedUrl}.png`,
-        body: buffer,
-        contentType,
-      });
-
-      if (error || !success) {
-        context.set.status = 500;
-        return { error: "Error uploading to S3" };
-      }
-
-      await prisma.generatedImage.create({
-        data: {
-          projectId: project.id,
-          urlPath: screenshotUrl,
-          imageUrl: `${Env.S3_ENDPOINT}/${encodedUrl}.png`,
-        },
-      });
-
-      polarClient.events.ingest({
-        events: [
-          {
-            name: "og-file-generated",
-            externalCustomerId: project?.user.id,
-            metadata: {
-              route: "/api/metered-route",
-              method: "GET",
-            },
-          },
-        ],
-      });
-
-      return { success: true };
-    }
-
-    return context.status(500, "Internal Server Error");
-  },
-  {
-    query: t.Object({
-      url: t.String(),
-    }),
-  }
-);
