+++ apps/server/src/modules/generate/screenshot/screenshot-route.ts
@@ -1,0 +1,15 @@
+import Elysia from "elysia";
+import { generateScreenshotService } from "./screenshot.service";
+import { ScreenshotQuerySchema } from "./screenshot-schema";
+
+export const screenshotModule = new Elysia().get(
+  "/generate-screenshot",
+  async (context) => {
+    const { url: queryUrl } = context.query;
+    const res = await generateScreenshotService(queryUrl);
+    return res;
+  },
+  {
+    query: ScreenshotQuerySchema,
+  }
+);
