+++ apps/server/src/modules/generate/screenshot/screenshot-schema.ts
@@ -1,0 +1,5 @@
+import { t } from "elysia";
+
+export const ScreenshotQuerySchema = t.Object({
+  url: t.String(),
+});
